Add files via upload
With scrounging here and there, I have managed to get this to work: Play a text, display the subtitles (with and without analysis), translate the utterance (on click), pause...
I would like to improve this in two ways: (1) I'd like for the file to stop playing and exit when the final utterance is done playing; (2) I would like the page to open displaying only the title of the text, and to start displaying the subtitles when PLAY is clicked.
My aim is to have a good template for the production of more such interactive listening files. I am not a coder, I am an "educator" who retired from teaching in 2007, trying to remain somewhat useful. Thank for any assistance you can give me.